+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes in keycloak-group-management will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0]
+
+### Added
+- Group admin can create a user group member based on username (REST API)
+
+### Fixed
+- Recalculate effective membership expires at when pending groups become active
+
 ## [1.0.4] - 2024-12-02
 
 ### Fixed